chau ObjectDAO y ObjectController
Hola NewsController, RSSFeedController, RSSFeedCategoryController,
NewsDAO, RSSFeedCategoryDAO, RSSFeedDAO!
