+++ local.properties
@@ -7,6 +7,6 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Mon Jun 27 14:38:28 GMT-03:00 2016
-ndk.dir=/Users/mbp/Library/Android/sdk/ndk-bundle
-sdk.dir=/Users/mbp/Library/Android/sdk
+#Mon Jun 27 19:51:30 GMT-03:00 2016
+ndk.dir=/Users/alejandrothomas/Library/Android/sdk/ndk-bundle
+sdk.dir=/Users/alejandrothomas/Library/Android/sdk
